Add Profiling Tools, Update dependencies, start work on standalone builds (#22)
* Experiment with standalone builds

* Update dependencies introduce anyhow for errors

* Better logging when loading openXR runtime

* Further reworks still not working

* Use costum branch of wgpu to relax validation on multiview

* Move certain logs to tracing level

* More stable + instrumentation

* Revert filtering changes on textures
