+++ src/engine.rs
@@ -2,6 +2,7 @@ use ash::vk;
 
 pub mod geometry;
 pub mod camera;
+pub mod texture;
 pub mod screen;
 pub mod entity;
 pub mod input;
@@ -13,10 +14,10 @@ pub const TARGET_VULKAN_VERSION: u32 = vk::make_api_version(0, 1, 1, 0);
 
 pub struct WgpuContext {
     pub vk_entry: ash::Entry,
-    pub vk_instance: ash::Instance,
-    pub vk_phys_device: ash::vk::PhysicalDevice,
+    pub vk_instance_ptr: u64,
+    pub vk_phys_device_ptr: u64,
+    pub vk_device_ptr: u64,
     pub queue_index: u32,
-    pub vk_device: ash::Device,
     pub instance: wgpu::Instance,
     pub device: wgpu::Device,
     pub physical_device: wgpu::Adapter,
@@ -24,7 +25,7 @@ pub struct WgpuContext {
 }
 
 pub trait WgpuLoader {
-    fn load_wgpu(&mut self) -> Option<WgpuContext>;
+    fn load_wgpu(&mut self) -> anyhow::Result<WgpuContext>;
 }
 
 pub trait WgpuRunner {
