+++ src/engine/vr.rs
@@ -1,16 +1,13 @@
-use std::error::Error;
-
-use ash::vk::{self, Handle};
+use anyhow::{Context, bail};
+use ash::vk::{self, Handle, QueueGlobalPriorityKHR};
 use hal::MemoryFlags;
 use openxr as xr;
-use wgpu::{
-    Device, Extent3d, Texture
-};
+use wgpu::{Device, Extent3d};
 use wgpu_hal as hal;
 
 use crate::conversions::vulkan_image_to_texture;
 
-use super::{WgpuLoader, WgpuRunner, TARGET_VULKAN_VERSION};
+use super::{WgpuLoader, WgpuRunner, TARGET_VULKAN_VERSION, texture::Texture2D};
 
 pub struct OpenXRContext {
     pub entry: openxr::Entry,
@@ -22,14 +19,29 @@ pub struct OpenXRContext {
 
 pub const VIEW_TYPE: openxr::ViewConfigurationType = openxr::ViewConfigurationType::PRIMARY_STEREO;
 pub const VIEW_COUNT: u32 = 2;
+pub const SWAPCHAIN_COLOR_FORMAT: wgpu::TextureFormat = wgpu::TextureFormat::Bgra8Unorm;
+pub const VK_SWAPCHAIN_COLOR_FORMAT: vk::Format = vk::Format::B8G8R8A8_SRGB;
 
-pub fn enable_xr_runtime() -> Result<OpenXRContext, Box<dyn Error>> {
+#[cfg(debug_assertions)] 
+pub fn openxr_layers() -> [&'static str; 0] {
+    [] //TODO: ["VK_LAYER_KHRONOS_validation"]
+}
+
+#[cfg(not(debug_assertions))]
+pub fn openxr_layers() -> [&'static str; 0] {
+    []
+}
+
+pub fn enable_xr_runtime() -> anyhow::Result<OpenXRContext> {
+    #[cfg(not(target_os = "android"))]
     let entry = openxr::Entry::linked();
+    #[cfg(target_os = "android")]
+    let entry = unsafe { openxr::Entry::load()? };
 
     #[cfg(target_os = "android")]
-    entry.initialize_android_loader().unwrap();
+    entry.initialize_android_loader()?;
 
-    let available_extensions = entry.enumerate_extensions().unwrap();
+    let available_extensions = entry.enumerate_extensions()?;
     log::info!("Available extensions: {:?}", available_extensions);
     assert!(available_extensions.khr_vulkan_enable2);
 
@@ -42,6 +54,7 @@ pub fn enable_xr_runtime() -> Result<OpenXRContext, Box<dyn Error>> {
     }
     log::info!("Enabled extensions: {:?}", enabled_extensions);
 
+    log::info!("Loading OpenXR Runtime...");
     let instance = entry.create_instance(
         &openxr::ApplicationInfo {
             application_name: "VR Screen Viewer",
@@ -50,7 +63,7 @@ pub fn enable_xr_runtime() -> Result<OpenXRContext, Box<dyn Error>> {
             engine_version: 0,
         },
         &enabled_extensions,
-        &[],
+        &openxr_layers(),
     )?;
 
     let props = instance.properties()?;
@@ -82,30 +95,42 @@ pub fn enable_xr_runtime() -> Result<OpenXRContext, Box<dyn Error>> {
     })
 }
 
+#[cfg(not(debug_assertions))]
+fn instance_flags() -> hal::InstanceFlags {
+    hal::InstanceFlags::empty()
+}
+
+#[cfg(debug_assertions)]
+fn instance_flags() -> hal::InstanceFlags {
+    hal::InstanceFlags::empty() | hal::InstanceFlags::VALIDATION | hal::InstanceFlags::DEBUG
+}
+
 impl WgpuLoader for OpenXRContext {
-    fn load_wgpu(&mut self) -> Option<super::WgpuContext> {
+    fn load_wgpu(&mut self) -> anyhow::Result<super::WgpuContext> {
         // OpenXR wants to ensure apps are using the correct graphics card and Vulkan features and
         // extensions, so the instance and device MUST be set up before Instance::create_session.
 
+        let wgpu_limits = wgpu::Limits::downlevel_webgl2_defaults();
+
+        let wgpu_features = wgpu::Features::MULTIVIEW;
         let vk_target_version = TARGET_VULKAN_VERSION; // Vulkan 1.1 guarantees multiview support
         let vk_target_version_xr = xr::Version::new(1, 1, 0);
 
         let reqs = self
             .instance
-            .graphics_requirements::<xr::Vulkan>(self.system)
-            .unwrap();
+            .graphics_requirements::<xr::Vulkan>(self.system)?;
 
         if vk_target_version_xr < reqs.min_api_version_supported
             || vk_target_version_xr.major() > reqs.max_api_version_supported.major()
         {
-            panic!(
+            bail!(
                 "OpenXR runtime requires Vulkan version > {}, < {}.0.0",
                 reqs.min_api_version_supported,
                 reqs.max_api_version_supported.major() + 1
             );
         }
 
-        let vk_entry = unsafe { ash::Entry::load().unwrap() };
+        let vk_entry = unsafe { ash::Entry::load()? };
         log::info!("Successfully loaded Vulkan entry");
 
         let vk_app_info = vk::ApplicationInfo::builder()
@@ -113,23 +138,14 @@ impl WgpuLoader for OpenXRContext {
             .engine_version(0)
             .api_version(vk_target_version);
 
-        let mut flags = hal::InstanceFlags::empty();
-        #[cfg(debug_assertions)]
-        {
-            flags |= hal::InstanceFlags::VALIDATION;
-            flags |= hal::InstanceFlags::DEBUG;
-        }
+        let flags = instance_flags();
 
-        let queue_index = 0;
         let instance_extensions =
-            <hal::api::Vulkan as hal::Api>::Instance::required_extensions(&vk_entry, flags)
-                .unwrap();
+            <hal::api::Vulkan as hal::Api>::Instance::required_extensions(&vk_entry, vk_target_version, flags)?;
 
         log::info!("Requested instance extensions: {:?}", instance_extensions);
-        let instance_extensions_ptrs = instance_extensions
-            .iter()
-            .map(|x| x.as_ptr())
-            .collect::<Vec<_>>();
+        let instance_extensions_ptrs: Vec<_> =
+            instance_extensions.iter().map(|x| x.as_ptr()).collect();
 
         let create_info = vk::InstanceCreateInfo::builder()
             .application_info(&vk_app_info)
@@ -143,9 +159,9 @@ impl WgpuLoader for OpenXRContext {
                     std::mem::transmute(vk_entry.static_fn().get_instance_proc_addr),
                     &create_info as *const _ as *const _,
                 )
-                .expect("XR error creating Vulkan instance")
+                .context("XR error creating Vulkan instance")?
                 .map_err(vk::Result::from_raw)
-                .expect("Vulkan error creating Vulkan instance");
+                .context("Vulkan error creating Vulkan instance")?;
             ash::Instance::load(
                 vk_entry.static_fn(),
                 vk::Instance::from_raw(vk_instance as _),
@@ -154,11 +170,10 @@ impl WgpuLoader for OpenXRContext {
 
         log::info!("Successfully created Vulkan instance");
 
-        let vk_physical_device = vk::PhysicalDevice::from_raw(
-            unsafe { self.instance
-                .vulkan_graphics_device(self.system, vk_instance.handle().as_raw() as _)
-                .unwrap() as _ },
-        );
+        let vk_physical_device = vk::PhysicalDevice::from_raw(unsafe {
+            self.instance
+                .vulkan_graphics_device(self.system, vk_instance.handle().as_raw() as _)? as _
+        });
 
         let vk_device_properties =
             unsafe { vk_instance.get_physical_device_properties(vk_physical_device) };
@@ -184,7 +199,7 @@ impl WgpuLoader for OpenXRContext {
                         None
                     }
                 })
-                .expect("Vulkan device has no graphics queue")
+                .context("Vulkan device has no graphics queue")?
         };
 
         log::info!("Got Vulkan queue family index {}", queue_family_index);
@@ -194,87 +209,90 @@ impl WgpuLoader for OpenXRContext {
                 vk_entry.clone(),
                 vk_instance.clone(),
                 vk_target_version,
-                0, //TODO: is this correct?
+                0,
                 instance_extensions,
                 flags,
-                false, //TODO: is this correct?
-                None,
-            )
-            .unwrap()
+                false,
+                Some(Box::new(())),
+            )?
         };
-        let hal_exposed_adapter = hal_instance.expose_adapter(vk_physical_device).unwrap();
+        let hal_exposed_adapter = hal_instance.expose_adapter(vk_physical_device).context("Cannot expose WGpu-Hal adapter")?;
 
         log::info!("Created WGPU-HAL instance and adapter");
-        
-        let device_descriptor = wgpu::DeviceDescriptor {
-            features: wgpu::Features::MULTIVIEW,
-            ..Default::default()
-        };
 
-        //TODO actually check if the extensions are available and avoid using then in the loaders
+        //TODO actually check if the extensions are available and avoid using them in the loaders
         let mut device_extensions = hal_exposed_adapter
             .adapter
-            .required_device_extensions(device_descriptor.features);
+            .required_device_extensions(wgpu_features);
 
         #[cfg(target_os = "windows")]
         device_extensions.push(ash::extensions::khr::ExternalMemoryWin32::name());
 
         log::info!("Requested device extensions: {:?}", device_extensions);
+        
+        let family_info = vk::DeviceQueueCreateInfo::builder()
+            .queue_family_index(queue_family_index)
+            .queue_priorities(&[1.0])
+            .push_next(&mut vk::DeviceQueueGlobalPriorityCreateInfoKHR::builder()
+                            .global_priority(QueueGlobalPriorityKHR::REALTIME_EXT)
+            )
+            .build();
 
         let device_extensions_ptrs = device_extensions
             .iter()
             .map(|x| x.as_ptr())
             .collect::<Vec<_>>();
 
-        let uab_types = hal::UpdateAfterBindTypes::from_limits(
-            &wgpu::Limits::default(),
-            &vk_device_properties.limits,
-        );
+        let mut enabled_features = hal_exposed_adapter
+            .adapter
+            .physical_device_features(&device_extensions, wgpu_features);
 
-        let vk_device = {
-            unsafe {
-                let vk_device = self
-                    .instance
-                    .create_vulkan_device(
-                        self.system,
-                        std::mem::transmute(vk_entry.static_fn().get_instance_proc_addr),
-                        vk_physical_device.as_raw() as _,
-                        &vk::DeviceCreateInfo::builder()
-                        .enabled_extension_names(&device_extensions_ptrs)
-                        .queue_create_infos(&[vk::DeviceQueueCreateInfo::builder()
-                            .queue_family_index(queue_family_index)
-                            .queue_priorities(&[1.0])
-                            .build()])
-                        .push_next(&mut vk::PhysicalDeviceMultiviewFeatures {
-                            multiview: vk::TRUE,
-                            ..Default::default()
-                        }) 
-                        as *const _ as *const _,
-                    )
-                    .expect("XR error creating Vulkan device")
-                    .map_err(vk::Result::from_raw)
-                    .expect("Vulkan error creating Vulkan device");
-                
-                log::info!("Creating ash vulkan device device from native");
-                ash::Device::load(vk_instance.fp_v1_0(), vk::Device::from_raw(vk_device as _))
-            }
-        };
+        let device_create_info = enabled_features
+            .add_to_device_create_builder(
+                vk::DeviceCreateInfo::builder()
+                    .enabled_extension_names(&device_extensions_ptrs)
+                    .queue_create_infos(&[family_info])
+                    .push_next(&mut vk::PhysicalDeviceMultiviewFeatures {
+                        multiview: vk::TRUE,
+                        ..Default::default()
+                    }),
+            )
+            .build();
 
+        let vk_device =
+            {
+                unsafe {
+                    let vk_device = self
+                        .instance
+                        .create_vulkan_device(
+                            self.system,
+                            std::mem::transmute(vk_entry.static_fn().get_instance_proc_addr),
+                            vk_physical_device.as_raw() as _,
+                            &device_create_info as *const _ as *const _,
+                        )
+                        .context("XR error creating Vulkan device")?
+                        .map_err(vk::Result::from_raw)
+                        .context("Vulkan error creating Vulkan device")?;
+
+                    log::info!("Creating ash vulkan device device from native");
+                    ash::Device::load(vk_instance.fp_v1_0(), vk::Device::from_raw(vk_device as _))
+                }
+            };
+
+        let vk_device_ptr = vk_device.handle().as_raw();
         log::info!("Successfully created Vulkan device");
 
         let hal_device = unsafe {
             hal_exposed_adapter
                 .adapter
                 .device_from_raw(
-                    vk_device.clone(),
+                    vk_device,
                     true, //    TODO: is this right?
                     &device_extensions,
-                    device_descriptor.features,
-                    uab_types,
-                    queue_family_index,
-                    queue_index,
-                )
-                .unwrap()
+                    wgpu_features,
+                    family_info.queue_family_index,
+                    0
+                )?
         };
 
         log::info!("Successfully created WGPU-HAL device from vulkan device");
@@ -284,22 +302,31 @@ impl WgpuLoader for OpenXRContext {
         let wgpu_adapter = unsafe { wgpu_instance.create_adapter_from_hal(hal_exposed_adapter) };
         let (wgpu_device, wgpu_queue) = unsafe {
             wgpu_adapter
-                .create_device_from_hal(hal_device, &device_descriptor, None)
-                .unwrap()
+                .create_device_from_hal(
+                    hal_device,
+                    &wgpu::DeviceDescriptor {
+                        features: wgpu_features,
+                        limits: wgpu_limits,
+                        label: None
+                    },
+                    None,
+                )?
         };
 
         log::info!("Successfully created WGPU context");
 
-        Some(super::WgpuContext {
+        log::info!("Queue timestamp period: {}", wgpu_queue.get_timestamp_period());
+
+        Ok(super::WgpuContext {
             instance: wgpu_instance,
             device: wgpu_device,
             physical_device: wgpu_adapter,
             queue: wgpu_queue,
-            queue_index: queue_family_index,
-            vk_device,
+            queue_index: family_info.queue_family_index,
             vk_entry,
-            vk_instance,
-            vk_phys_device: vk_physical_device,
+            vk_device_ptr,
+            vk_instance_ptr: vk_instance.handle().as_raw(),
+            vk_phys_device_ptr: vk_physical_device.as_raw(),
         })
     }
 }
@@ -315,24 +342,21 @@ impl OpenXRContext {
         &self,
         xr_session: &openxr::Session<openxr::Vulkan>,
         device: &Device,
-    ) -> (
+    ) -> anyhow::Result<(
         openxr::Swapchain<openxr::Vulkan>,
         vk::Extent2D,
-        Vec<Texture>,
-    ) {
+        Vec<Texture2D>,
+    )> {
         log::info!("Creating OpenXR swapchain");
 
         // Fetch the views we need to render to (the eye screens on the HMD)
         let views = self
             .instance
-            .enumerate_view_configuration_views(self.system, VIEW_TYPE)
-            .unwrap();
+            .enumerate_view_configuration_views(self.system, VIEW_TYPE)?;
         assert_eq!(views.len(), VIEW_COUNT as usize);
         assert_eq!(views[0], views[1]);
 
         // Create the OpenXR swapchain
-        let vk_color_format = vk::Format::B8G8R8A8_SRGB;
-        let color_format = wgpu::TextureFormat::Bgra8Unorm;
         let resolution = vk::Extent2D {
             width: views[0].recommended_image_rect_width,
             height: views[0].recommended_image_rect_height,
@@ -342,20 +366,18 @@ impl OpenXRContext {
                 create_flags: openxr::SwapchainCreateFlags::EMPTY,
                 usage_flags: openxr::SwapchainUsageFlags::COLOR_ATTACHMENT
                     | openxr::SwapchainUsageFlags::SAMPLED,
-                format: vk_color_format.as_raw() as _,
+                format: VK_SWAPCHAIN_COLOR_FORMAT.as_raw() as _,
                 sample_count: 1,
                 width: resolution.width,
                 height: resolution.height,
                 face_count: 1,
                 array_size: VIEW_COUNT,
                 mip_count: 1,
-            })
-            .unwrap();
+            })?;
 
         // Create image views for the swapchain
         let swapcain_textures: Vec<_> = xr_swapchain
-            .enumerate_images()
-            .unwrap()
+            .enumerate_images()?
             .into_iter()
             .map(|image| {
                 let wgpu_tex_desc = wgpu::TextureDescriptor {
@@ -368,10 +390,13 @@ impl OpenXRContext {
                     mip_level_count: 1,
                     sample_count: 1,
                     dimension: wgpu::TextureDimension::D2,
-                    format: color_format,
-                    usage: wgpu::TextureUsages::RENDER_ATTACHMENT | wgpu::TextureUsages::TEXTURE_BINDING,
+                    format: SWAPCHAIN_COLOR_FORMAT,
+                    view_formats:  &[],
+                    usage: wgpu::TextureUsages::RENDER_ATTACHMENT
+                        | wgpu::TextureUsages::TEXTURE_BINDING
+                        | wgpu::TextureUsages::COPY_DST,
                 };
-                
+
                 let wgpu_hal_tex_desc = wgpu_hal::TextureDescriptor {
                     label: wgpu_tex_desc.label,
                     size: wgpu_tex_desc.size,
@@ -379,20 +404,24 @@ impl OpenXRContext {
                     sample_count: wgpu_tex_desc.sample_count,
                     dimension: wgpu_tex_desc.dimension,
                     format: wgpu_tex_desc.format,
-                    usage: wgpu_hal::TextureUses::COLOR_TARGET | wgpu_hal::TextureUses::RESOURCE, //todo what here?
+                    view_formats:  vec!(),
+                    usage: wgpu_hal::TextureUses::COLOR_TARGET | wgpu_hal::TextureUses::COPY_DST,
                     memory_flags: MemoryFlags::empty(),
                 };
-                
+
                 // Create a WGPU image view for this image
-                vulkan_image_to_texture(
+                // TODO: Move this to Texture2D::from_vk_image
+                let wgpu_texture= vulkan_image_to_texture(
                     device,
                     vk::Image::from_raw(image),
                     wgpu_tex_desc,
                     wgpu_hal_tex_desc,
-                )
+                );
+
+                Texture2D::from_wgpu(device, wgpu_texture)
             })
             .collect();
 
-        (xr_swapchain, resolution, swapcain_textures)
+        Ok((xr_swapchain, resolution, swapcain_textures))
     }
 }
