+++ src/loaders.rs
@@ -1,10 +1,12 @@
-use std::error::Error;
 use wgpu::{Device, Instance};
 
+use crate::engine::texture::Texture2D;
+
+#[cfg(target_os="windows")]
 pub mod katanga_loader;
 
 pub struct TextureSource {
-    pub texture: wgpu::Texture,
+    pub texture: Texture2D,
     pub width: u32,
     pub height: u32,
     pub stereo_mode: StereoMode,
@@ -24,7 +26,7 @@ pub trait Loader {
         &mut self,
         instance: &Instance,
         device: &Device,
-    ) -> Result<TextureSource, Box<dyn Error>>;
+    ) -> anyhow::Result<TextureSource>;
 
     fn is_invalid(&self) -> bool;
 }
